fixing leave message handling, active beacons is a set
Leave message handling is changed to treat mobile beacons
differently from region beacons and gps regions. Should
fix a bug but still working on a testcase for it.

active beacons should be a set. you shouldn't end up
with multiple "active" entries for the same beacon. Let's
enforce that with the correct data structure.